ci: automatically bump the `phylum` formula
This changes adds a workflow for bumping the CLI version references in
the Formula and creating a PR with the changes. It is configured to be
triggered by repository dispatch events which come from outside of this
repository. It requires write access to the repository by providing a
personal access token (PAT) with `repo` scope. See the comments in the
workflow for more detail.

Closes #14