ci: checkout repo before setting up brew (#24)
With the `homebrew-cli` repo updated now to include a first draft of the "bump" workflow, I'm starting to interate on it...triggering the workflow locally for the CLI v4.7.0 release. The [first attempt failed](https://github.com/phylum-dev/homebrew-cli/actions/runs/4287457567/jobs/7468336703) but it is not exactly clear how/why it failed...at least from reading the logs there. This PR takes the assumption that maybe the `actions/checkout` action needs to occur before the `Homebrew/actions/setup-homebrew` action. It also enables the debug output for that latter action.