ci: dump homebrew actions in favor of manual formula bumping
The `Homebrew/actions/*` actions do not appear to be flexible enough to work in the manner intended for this repository. They make certain assumptions that do not hold. Specifically, the `brew bump-formula-pr` command attempts to `git push` in a way that does not work with the `phylum-bot` user token and signing restrictions in place. Several attempts were made to work around these limitations, but the time taken to figure that out was dragging on and a decision was made to just re-write the workflow to manually swap out the `phylum` formula values for `url` and `sha256`, commit the changes, push them, and start a PR. The changes made here adhere to `actionlint` and `shellcheck` and are believed to work as intended...but it won't be truly possible to tell until the workflow is triggered after it gets merged to the `main` branch.