ci: address PR comments
The `continue-on-error` logic was replicated from the `cli` and `phylum-ci` workflows for automatically updating dependencies. It was used there because there is a chance that there are NO dependencies that get updated...which would cause the `git` commands in that step to fail. `continue-on-error` is used for that situation so that the overall job does not show as a failure. In the case of this workflow, the files are explicitly modified with (presumably) a new/different CLI version and artifact SHA. The `curl -O` option was used to simplify that command. The text of the PR (the body) was updated to be more verbose, explaining in detail the process for reviewing and merging the PR.