ci: conditionally set the environment with matrix variable (#38)
This change attempts to expand the matrix to include a `self_hosted`
property only for the `os` entries that are indeed self-hosted. Since
the property won't exist for some of the `os` entries, it should follow
the behavior described [on the GitHub Contexts
documentation](https://docs.github.com/en/actions/learn-github-actions/contexts):

> If you attempt to dereference a non-existent property, it will
evaluate to an empty string.