ci: bypass `publish` workflow for disabled formula
The `publish` workflow will no longer work now that the `phylum` formula
has been disabled. This is because the workflow depends on the formula
to be installed and the whole point of disabling the formula is so that
it can **not** be installed. It turns out that it doesn't matter if the
rest of the contents of the formula are incorrect so long as the
`disable!` statement is included along with the first block of metadata,
to include the new version.