Disable the `phylum` formula (#65)
This change disables the `phylum` formula for the v6.1.2 release of the
Phylum CLI, effective 22 FEB 2024. Users attempting to upgrade/install
the formula will see the following:

```
❯ brew install phylum
Error: phylum-dev/cli/phylum has been disabled because it is available as `phylum-cli` in core tap!
To migrate:

  brew uninstall phylum
  brew untap phylum-dev/cli
  brew install phylum-cli

!
```

The `publish` workflow will no longer work now that the `phylum` formula
has been disabled. This is because the workflow depends on the formula
to be installed and the whole point of disabling the formula is so that
it can **not** be installed. It turns out that it doesn't matter if the
rest of the contents of the formula are incorrect so long as the
`disable!` statement is included along with the first block of metadata,
to include the new version.

Therefore, in order to allow this PR to complete, the `publish` workflow
was deleted so it would not be triggered. This should not matter going
forward as the repository is slated to be archived and possibly even
deleted.